feat: remove keepassxc from scratchpads
The browser integration feature of keepassxc doesn't seem to work well
with scratchpads. The pop-up dialog for auto-typing username and
password doesn't show up in the current window, it spawns inside of the
scratchpad instead which is hidden from view until the keyboard shortcut
for accessing the scratchpad is pressed.